+++ next-docs/components/ManifestCard.tsx
@@ -13,13 +13,13 @@ const ManifestCard: React.FC<Props> = ({
 }) => (
   <div
     className={classNames(
-      'moon-light flex flex-col gap-4 py-6 px-8 rounded-moon-s-lg',
+      'theme-moon-light flex flex-col gap-4 py-6 px-8 rounded-moon-s-lg',
       isHighlighted ? 'bg-frieza text-goten' : 'bg-gohan text-bulma'
     )}
   >
     <h3 className="text-moon-24 font-medium">{title}</h3>
     {!!contentItems.length && (
-      <ul className="list-disc ml-6">
+      <ul className="list-disc ms-6">
         {contentItems.map((item: string) => (
           <li className="text-moon-16">{item}</li>
         ))}