+++ next-docs/components/settings/common/SettingsButton.tsx
@@ -13,7 +13,7 @@ const SettingsButton: React.FC<Props> = ({ onClick, isRtlEnabled }) => (
     aria-pressed="false"
     className={classNames(
       'text-black bg-white hover:bg-gray-200 fixed bottom-4 inline-flex flex-shrink-0 border-2 border-transparent rounded-full cursor-pointer transition-colors ease-in-out duration-200 focus:outline-none z-50 shadow-md',
-      isRtlEnabled ? 'left-4' : 'right-4 '
+      isRtlEnabled ? 'left-4' : 'right-4'
     )}
   >
     <MediaTuner className="text-moon-32" />