+++ next-docs/public/examples/checkboxTW/WithLabel.tsx
@@ -1,0 +1,10 @@
+import { Checkbox } from '@heathmont/moon-core-tw';
+
+const Example = () => (
+  <Checkbox
+    label="I agree to receive bonus & marketing emails."
+    id="checkbox1"
+  />
+);
+
+export default Example;