+++ workspaces/core/src/button/private/buttonSizes/xsPadding.ts
@@ -2,10 +2,10 @@ import type ButtonProps from '../types/ButtonProps';
 
 const xsPadding = ({ icon, iconLeft, iconRight, iconOnly }: ButtonProps) => {
   if (icon === 'left' || iconLeft) {
-    return 'p-1 pr-2';
+    return 'p-1 pe-2';
   }
   if (icon === 'right' || iconRight) {
-    return 'p-1 pl-2';
+    return 'p-1 ps-2';
   }
   if (icon === 'only' || iconOnly) {
     return 'p-1';