+++ workspaces/core/src/label/utils/setPadding.ts
@@ -2,7 +2,7 @@ import LabelProps from '../private/types/LabelProps';
 
 const setPadding = ({ size, iconRight, iconLeft }: LabelProps) =>
   `${size === '2xs' ? 'py-0' : 'py-1'} 
-  ${iconRight ? 'pr-1' : 'pr-2'} 
-  ${iconLeft ? 'pl-1' : 'pl-2'}`;
+  ${iconRight ? 'pe-1' : 'pe-2'} 
+  ${iconLeft ? 'ps-1' : 'ps-2'}`;
 
 export default setPadding;