+++ packages/popover/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 9.1.5
+
+### Patch Changes
+
+- vip templates
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@9.1.5
+  - @heathmont/moon-components@9.1.5
+  - @heathmont/moon-themes@9.1.5
+  - @heathmont/moon-utils@9.1.5
+
+## 9.1.4
+
+### Patch Changes
+
+- Casino template
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@9.1.4
+  - @heathmont/moon-components@9.1.4
+  - @heathmont/moon-themes@9.1.4
+  - @heathmont/moon-utils@9.1.4
+
 ## 9.1.3
 
 ### Patch Changes