+++ workspaces/icons/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 9.1.5
+
+### Patch Changes
+
+- vip templates
+
+## 9.1.4
+
+### Patch Changes
+
+- Casino template
+
 ## 9.1.3
 
 ### Patch Changes